+++ src/dotnet-interactive/Startup.cs
@@ -4,7 +4,7 @@
 using System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.DotNet.Interactive.App.CommandLine;
-using Microsoft.DotNet.Interactive.App.HttpRouting;
+using Microsoft.DotNet.Interactive.App.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.FileProviders;
